Relocated LazyLoadingSupportInterface and fixed upgrade notes
