+++ models/DataObject/ClassDefinition/Data/LazyLoadingSupportInterface.php
@@ -15,7 +15,7 @@
  * @license    http://www.pimcore.org/license     GPLv3 and PEL
  */
 
-namespace Pimcore\Model\DataObject;
+namespace Pimcore\Model\DataObject\ClassDefinition\Data;
 
 /**
  * Marker interface for data types that support lazy loading
