+++ models/DataObject/Fieldcollection/Data/AbstractData.php
@@ -19,6 +19,7 @@ namespace Pimcore\Model\DataObject\Fieldcollection\Data;
 
 use Pimcore\Model;
 use Pimcore\Model\DataObject\Concrete;
+use Pimcore\Model\DataObject\ClassDefinition\Data\LazyLoadingSupportInterface;
 
 /**
  * @method Dao getDao()
@@ -168,7 +169,7 @@ abstract class AbstractData extends Model\AbstractModel implements Model\DataObj
         $lazyLoadedFieldNames = [];
         $fields = $this->getDefinition()->getFieldDefinitions(['suppressEnrichment' => true]);
         foreach ($fields as $field) {
-            if (($field instanceof  Model\DataObject\LazyLoadingSupportInterface || method_exists($field, 'getLazyLoading'))
+            if (($field instanceof LazyLoadingSupportInterface || method_exists($field, 'getLazyLoading'))
                             && $field->getLazyLoading()) {
                 $lazyLoadedFieldNames[] = $field->getName();
             }
