+++ models/DataObject/Objectbrick/Data/AbstractData.php
@@ -21,6 +21,7 @@ use Pimcore\Model;
 use Pimcore\Model\DataObject;
 use Pimcore\Model\DataObject\Concrete;
 use Pimcore\Model\DataObject\Exception\InheritanceParentNotFoundException;
+use Pimcore\Model\DataObject\ClassDefinition\Data\LazyLoadingSupportInterface;
 
 /**
  * @method Dao getDao()
@@ -254,7 +255,7 @@ abstract class AbstractData extends Model\AbstractModel implements Model\DataObj
         $lazyLoadedFieldNames = [];
         $fields = $this->getDefinition()->getFieldDefinitions(['suppressEnrichment' => true]);
         foreach ($fields as $field) {
-            if (($field instanceof  DataObject\LazyLoadingSupportInterface || method_exists($field, 'getLazyLoading'))
+            if (($field instanceof  LazyLoadingSupportInterface || method_exists($field, 'getLazyLoading'))
                             && $field->getLazyLoading()) {
                 $lazyLoadedFieldNames[] = $field->getName();
             }
